requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg2==2.9.9
GitPython==3.1.43
langchain-openai==0.2.1
openai==1.51.0